requirements.txt
discord.py==2.6.0
python-dotenv==1.1.1
requests==2.32.5
